+++ src/deploy/cdk/constructs/index.ts
@@ -1,1 +1,4 @@
-// TODO - we'll likely have constructs for common Komodo infra like DynamoDB, S3, Fargate services, etc.
+export * from './fargatePublicService.js';
+export * from './fargatePrivateService.js';
+export * from './wafWebAcl.js';
+export * from './metricFilterAlarm.js';